Move AWS and Upload out of server.ts
