+++ daos/AwsDaos.ts
@@ -1,11 +1,9 @@
-import AWS from 'aws-sdk';
-import dotenv from "dotenv";
+// Configures and connects to AWS S3 image storage.
 
-dotenv.config();
+import AWS from "aws-sdk";
+
 export const s3 = new AWS.S3({
     accessKeyId: process.env.AWS_S3_KEY_ID,
     secretAccessKey: process.env.AWS_S3_KEY,
     region: process.env.AWS_REGION
 });
-
-export default s3;
