+++ package.json
@@ -25,6 +25,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
+    "heroku": "^7.67.1",
     "mocha": "^10.1.0",
     "mongoose": "^6.2.1",
     "multer": "^1.4.5-lts.1",
